requirements.txt
# AI-Tools Python Dependencies
# Install with: uv pip install -r requirements.txt

# Core AI Providers
openai
anthropic
google-genai

# LangChain (optional, for advanced chains)
langchain
langchain-openai
langchain-anthropic

# CLI Tools
shell-gpt
rich
typer

# Web Framework
fastapi
uvicorn[standard]
jinja2
httpx

# ML/Deep Learning (optional, for local inference)
# torch
# transformers
# accelerate

# Utilities
requests
python-dotenv
pyyaml
docker

# Performance (Linux only)
# uvloop  # Not supported on Windows
# httptools
